+++ urbanairship-automation/src/main/java/com/urbanairship/automation/InAppAutomation.java
@@ -83,6 +83,7 @@ public class InAppAutomation extends AirshipComponent implements InAppAutomation
 
     private final AtomicBoolean isStarted = new AtomicBoolean(false);
     private Subscription subscription;
+    private final AirshipRuntimeConfig config;
 
     private final AutomationDriver driver = new AutomationDriver() {
         @Override
@@ -181,11 +182,13 @@ public class InAppAutomation extends AirshipComponent implements InAppAutomation
         this.actionScheduleDelegate = new ActionsScheduleDelegate();
         this.inAppMessageScheduleDelegate = new InAppMessageScheduleDelegate(inAppMessageManager);
         this.frequencyLimitManager = new FrequencyLimitManager(context, runtimeConfig);
+        this.config = runtimeConfig;
     }
 
     @VisibleForTesting
     InAppAutomation(@NonNull Context context,
                     @NonNull PreferenceDataStore preferenceDataStore,
+                    @NonNull AirshipRuntimeConfig runtimeConfig,
                     @NonNull PrivacyManager privacyManager,
                     @NonNull AutomationEngine engine,
                     @NonNull AirshipChannel airshipChannel,
@@ -209,6 +212,7 @@ public class InAppAutomation extends AirshipComponent implements InAppAutomation
         this.actionScheduleDelegate = actionsScheduleDelegate;
         this.inAppMessageScheduleDelegate = inAppMessageScheduleDelegate;
         this.frequencyLimitManager = frequencyLimitManager;
+        this.config = runtimeConfig;
     }
 
     /**
@@ -219,6 +223,9 @@ public class InAppAutomation extends AirshipComponent implements InAppAutomation
     protected void init() {
         super.init();
 
+        if (config.getConfigOptions().autoPauseInAppAutomationOnLaunch) {
+            this.setPaused(true);
+        }
         this.automationEngine.setScheduleListener(new AutomationEngine.ScheduleListener() {
             @Override
             public void onScheduleExpired(@NonNull final Schedule<? extends ScheduleData> schedule) {