+++ CHANGELOG.md
@@ -2,12 +2,19 @@
 
 [Migration Guides](https://github.com/urbanairship/android-library/tree/main/documentation/migration)
 
+## Version 16.11.0, June 12, 2023
+Minor release that adds new config option `autoPauseInAppAutomationOnLaunch` to always pause IAA during app launch. 
+
+### Changes
+- Added `autoPauseInAppAutomationOnLaunch` config option
+
+
 ## Version 16.10.0, June 2, 2023
 
 Minor release that adds support for Android Live Updates, which bring functionality similar to iOS Live Activities to the Android platform.
 Live Updates make it easier to keep information updated in real time instead of receiving multiple notifications from the same app for things like a game’s latest score, food delivery status, or rideshare arrivals.
 
-## Changes
+### Changes
 - Added new `urbanairship-live-update` module.
 
 ## Version 16.9.3, May 24, 2023